The map class handles everything connected to the "word" and environment. It initializes as much as possible before the application starts (most importantly, the pathfinder "PathFinder"), this way the application can run smoothly, while just using a little more random access memory...
The Map class saves / initializes the game world by reading the JSON file, afterwards sprites are being saved ("SaveSprites()"), layers are being saved ("SaveLayers()") and areas are being saved ("SaveAreas()").

Last but not least: there are a few draw methods that are specialized in drawing specific layers, this way it i s far more easier to change the draw order mid project (for example when a Sim should be able to stand in front of a wall, instead of always on the backside, hence the wall would be overlapping the Sim 100% of the time then, which is not what you want). 

The "getImageOfCanvas()" method returns a snapshot of the canvas (all written seperated objects are being grouped and saved). This way the entire map can be drawn over and over by only drawing 1 image, instead of 18 (amount of layers) multiplied with 100 tiles counting width * 120 tiles counting height, this way we are saving 215.999 draw operations! This turned out to be a performance saver for the application.

There is also a basic event handler, if key input == x then do this. For example if key = l then show debug. 

For any questions regarding this class, pm / call / mail me.